add functionality for scanning OID matchers in a CT log entry
