+++ pkg/ct/identity.go
@@ -1,0 +1,67 @@
+// Copyright 2024 The Sigstore Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// This file copies some of the functionality in pkg/identity/identity.go
+// related to retrieving OID extension values and matching on them,
+// but refactors them to use the Google-specific fork of encoding/asn1 and crypto/x509.
+
+package ct
+
+import (
+	"encoding/asn1"
+	"fmt"
+
+	google_asn1 "github.com/google/certificate-transparency-go/asn1"
+	google_x509 "github.com/google/certificate-transparency-go/x509"
+)
+
+// getExtension gets a certificate extension by OID where the extension value is an
+// ASN.1-encoded string
+func getExtension(cert *google_x509.Certificate, oid google_asn1.ObjectIdentifier) (string, error) {
+	for _, ext := range cert.Extensions {
+		if !ext.Id.Equal(oid) {
+			continue
+		}
+		var extValue string
+		rest, err := asn1.Unmarshal(ext.Value, &extValue)
+		if err != nil {
+			return "", fmt.Errorf("%w", err)
+		}
+		if len(rest) != 0 {
+			return "", fmt.Errorf("unmarshalling extension had rest for oid %v", oid)
+		}
+		return extValue, nil
+	}
+	return "", nil
+}
+
+// OIDMatchesPolicy returns if a certificate contains both a given OID field and a matching value associated with that field
+// if true, it returns the OID extension and extension value that were matched on
+func OIDMatchesPolicy(cert *google_x509.Certificate, oid google_asn1.ObjectIdentifier, extensionValues []string) (bool, google_asn1.ObjectIdentifier, string, error) {
+	extValue, err := getExtension(cert, oid)
+	if err != nil {
+		return false, nil, "", fmt.Errorf("error getting extension value: %w", err)
+	}
+	if extValue == "" {
+		return false, nil, "", nil
+	}
+
+	for _, extensionValue := range extensionValues {
+		if extValue == extensionValue {
+			return true, oid, extValue, nil
+		}
+	}
+
+	return false, nil, "", nil
+}
