+++ pkg/ct/monitor_test.go
@@ -15,12 +15,16 @@
 package ct
 
 import (
+	"encoding/asn1"
 	"reflect"
 	"testing"
+
+	google_asn1 "github.com/google/certificate-transparency-go/asn1"
 
 	ct "github.com/google/certificate-transparency-go"
 	"github.com/google/certificate-transparency-go/x509"
 	"github.com/google/certificate-transparency-go/x509/pkix"
+	"github.com/sigstore/rekor-monitor/pkg/fulcio/extensions"
 	"github.com/sigstore/rekor-monitor/pkg/identity"
 )
 
@@ -29,7 +33,7 @@ const (
 	organizationName = "test-org"
 )
 
-func TestScanEntrySubject(t *testing.T) {
+func TestScanEntryCertSubject(t *testing.T) {
 	testCases := map[string]struct {
 		inputEntry    ct.LogEntry
 		inputSubjects []string
@@ -68,9 +72,74 @@ func TestScanEntrySubject(t *testing.T) {
 	}
 
 	for _, tc := range testCases {
-		logEntries, err := ScanEntrySubject(tc.inputEntry, tc.inputSubjects)
+		logEntries, err := ScanEntryCertSubject(tc.inputEntry, tc.inputSubjects)
 		if err != nil {
 			t.Errorf("received error scanning entry for subjects: %v", err)
+		}
+		expected := tc.expected
+		if logEntries == nil {
+			if expected != nil {
+				t.Errorf("received nil, expected log entry")
+			}
+		} else {
+			if !reflect.DeepEqual(logEntries, expected) {
+				t.Errorf("expected %v, received %v", expected, logEntries)
+			}
+		}
+	}
+}
+
+func TestScanEntryOIDExtensions(t *testing.T) {
+	cert, err := mockCertificateWithExtension(google_asn1.ObjectIdentifier{2, 5, 29, 17}, "test cert value")
+	if err != nil {
+		t.Errorf("Expected nil got %v", err)
+	}
+	unmatchedAsn1OID := asn1.ObjectIdentifier{2}
+	matchedAsn1OID := asn1.ObjectIdentifier{2, 5, 29, 17}
+	extValueString := "test cert value"
+	testCases := map[string]struct {
+		inputEntry         ct.LogEntry
+		inputOIDExtensions []extensions.OIDExtension
+		expected           []*identity.LogEntry
+	}{
+		"no matching subject": {
+			inputEntry: ct.LogEntry{
+				Index:    1,
+				X509Cert: cert,
+			},
+			inputOIDExtensions: []extensions.OIDExtension{
+				{
+					ObjectIdentifier: unmatchedAsn1OID,
+					ExtensionValues:  []string{extValueString},
+				},
+			},
+			expected: []*identity.LogEntry{},
+		},
+		"matching subject": {
+			inputEntry: ct.LogEntry{
+				Index:    1,
+				X509Cert: cert,
+			},
+			inputOIDExtensions: []extensions.OIDExtension{
+				{
+					ObjectIdentifier: matchedAsn1OID,
+					ExtensionValues:  []string{extValueString},
+				},
+			},
+			expected: []*identity.LogEntry{
+				{
+					Index:          1,
+					OIDExtension:   matchedAsn1OID,
+					ExtensionValue: extValueString,
+				},
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		logEntries, err := ScanEntryOIDExtensions(tc.inputEntry, tc.inputOIDExtensions)
+		if err != nil {
+			t.Errorf("received error scanning entry for oid extensions: %v", err)
 		}
 		expected := tc.expected
 		if logEntries == nil {
