+++ pkg/rekor/identity.go
@@ -18,7 +18,6 @@ import (
 	"bytes"
 	"context"
 	"crypto/x509"
-	"encoding/asn1"
 	"encoding/base64"
 	"errors"
 	"fmt"
@@ -34,7 +33,6 @@ import (
 	"github.com/sigstore/rekor/pkg/pki"
 	"github.com/sigstore/rekor/pkg/types"
 	"github.com/sigstore/rekor/pkg/util"
-	"github.com/sigstore/sigstore/pkg/cryptoutils"
 
 	// required imports to call init methods
 	_ "github.com/sigstore/rekor/pkg/types/alpine/v0.0.1"
@@ -49,11 +47,6 @@ import (
 	_ "github.com/sigstore/rekor/pkg/types/rfc3161/v0.0.1"
 	_ "github.com/sigstore/rekor/pkg/types/rpm/v0.0.1"
 	_ "github.com/sigstore/rekor/pkg/types/tuf/v0.0.1"
-)
-
-var (
-	certExtensionOIDCIssuer   = asn1.ObjectIdentifier{1, 3, 6, 1, 4, 1, 57264, 1, 1}
-	certExtensionOIDCIssuerV2 = asn1.ObjectIdentifier{1, 3, 6, 1, 4, 1, 57264, 1, 8}
 )
 
 func MatchLogEntryFingerprints(logEntryAnon models.LogEntryAnon, uuid string, entryFingerprints []string, monitoredFingerprints []string) []identity.LogEntry {
@@ -76,7 +69,7 @@ func MatchLogEntryCertificateIdentities(logEntryAnon models.LogEntryAnon, uuid s
 	matchedEntries := []identity.LogEntry{}
 	for _, monitoredCertID := range monitoredCertIDs {
 		for _, cert := range entryCertificates {
-			match, sub, iss, err := certMatchesPolicy(cert, monitoredCertID.CertSubject, monitoredCertID.Issuers)
+			match, sub, iss, err := identity.CertMatchesPolicy(cert, monitoredCertID.CertSubject, monitoredCertID.Issuers)
 			if err != nil {
 				return nil, fmt.Errorf("error with policy matching for UUID %s at index %d: %w", uuid, logEntryAnon.LogIndex, err)
 			} else if match {
@@ -116,7 +109,7 @@ func MatchLogEntryOIDs(logEntryAnon models.LogEntryAnon, uuid string, entryCerti
 	matchedEntries := []identity.LogEntry{}
 	for _, monitoredOID := range monitoredOIDMatchers {
 		for _, cert := range entryCertificates {
-			match, oid, extValue, err := oidMatchesPolicy(cert, monitoredOID.ObjectIdentifier, monitoredOID.ExtensionValues)
+			match, oid, extValue, err := identity.OIDMatchesPolicy(cert, monitoredOID.ObjectIdentifier, monitoredOID.ExtensionValues)
 			if err != nil {
 				return nil, fmt.Errorf("error with policy matching for UUID %s at index %d: %w", uuid, logEntryAnon.LogIndex, err)
 			}
@@ -275,101 +268,6 @@ func extractAllIdentities(verifiers []pki.PublicKey) ([]string, []*x509.Certific
 		}
 	}
 	return subjects, certificates, fps, nil
-}
-
-// getExtension gets a certificate extension by OID where the extension value is an
-// ASN.1-encoded string
-func getExtension(cert *x509.Certificate, oid asn1.ObjectIdentifier) (string, error) {
-	for _, ext := range cert.Extensions {
-		if !ext.Id.Equal(oid) {
-			continue
-		}
-		var extValue string
-		rest, err := asn1.Unmarshal(ext.Value, &extValue)
-		if err != nil {
-			return "", fmt.Errorf("%w", err)
-		}
-		if len(rest) != 0 {
-			return "", fmt.Errorf("unmarshalling extension had rest for oid %v", oid)
-		}
-		return extValue, nil
-	}
-	return "", nil
-}
-
-// getDeprecatedExtension gets a certificate extension by OID where the extension value is a raw string
-func getDeprecatedExtension(cert *x509.Certificate, oid asn1.ObjectIdentifier) (string, error) {
-	for _, ext := range cert.Extensions {
-		if ext.Id.Equal(oid) {
-			return string(ext.Value), nil
-		}
-	}
-	return "", nil
-}
-
-// certMatchesPolicy returns true if a certificate contains a given subject and optionally a given issuer
-// expectedSub and expectedIssuers can be regular expressions
-// certMatchesPolicy also returns the matched subject and issuer on success
-func certMatchesPolicy(cert *x509.Certificate, expectedSub string, expectedIssuers []string) (bool, string, string, error) {
-	sans := cryptoutils.GetSubjectAlternateNames(cert)
-	var issuer string
-	var err error
-	issuer, err = getExtension(cert, certExtensionOIDCIssuerV2)
-	if err != nil || issuer == "" {
-		// fallback to deprecated issuer extension
-		issuer, err = getDeprecatedExtension(cert, certExtensionOIDCIssuer)
-		if err != nil || issuer == "" {
-			return false, "", "", err
-		}
-	}
-	subjectMatches := false
-	regex, err := regexp.Compile(expectedSub)
-	if err != nil {
-		return false, "", "", fmt.Errorf("malformed subject regex: %w", err)
-	}
-	matchedSub := ""
-	for _, sub := range sans {
-		if regex.MatchString(sub) {
-			subjectMatches = true
-			matchedSub = sub
-		}
-	}
-	// allow any issuer
-	if len(expectedIssuers) == 0 {
-		return subjectMatches, matchedSub, issuer, nil
-	}
-
-	issuerMatches := false
-	for _, expectedIss := range expectedIssuers {
-		regex, err := regexp.Compile(expectedIss)
-		if err != nil {
-			return false, "", "", fmt.Errorf("malformed issuer regex: %w", err)
-		}
-		if regex.MatchString(issuer) {
-			issuerMatches = true
-		}
-	}
-	return subjectMatches && issuerMatches, matchedSub, issuer, nil
-}
-
-// oidMatchesPolicy returns if a certificate contains both a given OID field and a matching value associated with that field
-// if true, it returns the OID extension and extension value that were matched on
-func oidMatchesPolicy(cert *x509.Certificate, oid asn1.ObjectIdentifier, extensionValues []string) (bool, asn1.ObjectIdentifier, string, error) {
-	extValue, err := getExtension(cert, oid)
-	if err != nil {
-		return false, nil, "", fmt.Errorf("error getting extension value: %w", err)
-	}
-	if extValue == "" {
-		return false, nil, "", nil
-	}
-
-	for _, extensionValue := range extensionValues {
-		if extValue == extensionValue {
-			return true, oid, extValue, nil
-		}
-	}
-
-	return false, nil, "", nil
 }
 
 func GetPrevCurrentCheckpoints(client *client.Rekor, logInfoFile string) (*util.SignedCheckpoint, *util.SignedCheckpoint, error) {
